Composer update | Responsive mejorado Parametros y Usuarios

+++ app/Livewire/Dashboard/ParametrosComponent.php
@@ -24,7 +24,7 @@ class ParametrosComponent extends Component
     public function render()
     {
         $parametros = Parametro::buscar($this->keyword)
-            ->orderBy('updated_at', 'ASC')
+            ->orderBy('created_at', 'DESC')
             ->limit($this->rows)
             ->get();
 
@@ -54,6 +54,7 @@ class ParametrosComponent extends Component
         $this->reset([
             'parametro_id', 'nombre', 'tabla_id', 'valor', 'view', 'keyword'
         ]);
+        $this->resetErrorBag();
     }
 
     protected function rules($id = null)
@@ -67,8 +68,6 @@ class ParametrosComponent extends Component
 
     public function save()
     {
-        $type = 'success';
-        $message = 'Hola Mundo';
 
         $this->validate($this->rules($this->parametro_id));
 
@@ -82,27 +81,32 @@ class ParametrosComponent extends Component
             $message = "Parametro Actualizado";
         }
 
-        $parametro->nombre = $this->nombre;
-        if (!empty($this->tabla_id)){
-            $parametro->tabla_id = $this->tabla_id;
-        }
-        if (!empty($this->valor)){
-            $parametro->valor = $this->valor;
-        }
-        $parametro->save();
+        if ($parametro){
+            $parametro->nombre = $this->nombre;
+            if (!empty($this->tabla_id)){
+                $parametro->tabla_id = $this->tabla_id;
+            }
+            if (!empty($this->valor)){
+                $parametro->valor = $this->valor;
+            }
+            $parametro->save();
 
-        $this->alert($type, $message);
+            $this->alert('success', $message);
+        }
         $this->limpiar();
+        $this->dispatch('cerrarModal');
     }
 
     public function edit($id)
     {
         $parametro = Parametro::find($id);
-        $this->parametro_id = $parametro->id;
-        $this->nombre = $parametro->nombre;
-        $this->tabla_id = $parametro->tabla_id;
-        $this->valor = $parametro->valor;
-        $this->view = "edit";
+        if ($parametro){
+            $this->parametro_id = $parametro->id;
+            $this->nombre = $parametro->nombre;
+            $this->tabla_id = $parametro->tabla_id;
+            $this->valor = $parametro->valor;
+            $this->view = "edit";
+        }
     }
 
     #[On('buscar')]
@@ -129,12 +133,17 @@ class ParametrosComponent extends Component
     public function confirmed()
     {
         $parametro = Parametro::find($this->parametro_id);
-        $parametro->delete();
-        $this->limpiar();
-        $this->alert(
-            'success',
-            'Parametro Eliminado'
-        );
+        if ($parametro){
+            $parametro->delete();
+            $this->limpiar();
+            $this->alert('success', 'Parametro Eliminado.');
+        }
+    }
+
+    #[On('cerrarModal')]
+    public function cerrarModal()
+    {
+        //JS
     }
 
 }
